Extract an interface for ScannerEngineFacade
