+++ lib/src/main/java/org/sonarsource/scanner/lib/internal/facade/package-info.java
@@ -17,28 +17,5 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.scanner.lib.internal.facade.forked;
-
-import java.util.Map;
-import javax.annotation.Nullable;
-import org.sonarsource.scanner.lib.internal.facade.AbstractScannerEngineFacade;
-
-public class NewScannerEngineFacade extends AbstractScannerEngineFacade {
-  private final ScannerEngineLauncher launcher;
-
-  public NewScannerEngineFacade(Map<String, String> bootstrapProperties, ScannerEngineLauncher launcher,
-    boolean isSonarCloud, @Nullable String serverVersion) {
-    super(bootstrapProperties, isSonarCloud, serverVersion, launcher.isEngineCacheHit(), launcher.getJreCacheHit());
-    this.launcher = launcher;
-  }
-
-  @Override
-  protected boolean doAnalyze(Map<String, String> allProps) {
-    return launcher.execute(allProps);
-  }
-
-  @Override
-  public void close() throws Exception {
-    // nothing to do
-  }
-}
+@javax.annotation.ParametersAreNonnullByDefault
+package org.sonarsource.scanner.lib.internal.facade;
